Time: 37 ms (68.68%) | Memory: 66.9 MB (83.00%) - LeetSync

+++ 1014-k-closest-points-to-origin/k-closest-points-to-origin.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     vector<vector<int>> kClosest(vector<vector<int>>& points, int k) {
-        priority_queue<pair<double,int>>pq;
+        priority_queue<pair<float,int>>pq;
         for(int i = 0;i<points.size();i++){
-            double dis = sqrt((points[i][0]*points[i][0]) + (points[i][1]*points[i][1]));
+            float dis = sqrt((points[i][0]*points[i][0]) + (points[i][1]*points[i][1]));
             pq.push({dis,i});
             if(pq.size()>k)pq.pop();
         }
